requirements.txt
beautifulsoup4==4.7.0
GitPython==2.1.11
html5lib==1.0.1
urwid==2.0.1
